#12 fixed abstraction of the side menu and app bar, fixed the position of workspace component on the app bar
